feat: Place order fixes, wWsJsonClient as interface
