+++ package.json
@@ -6,6 +6,7 @@
   "types": "dist/web.d.ts",
   "exports": {
     "./wsJsonClient": "./dist/client/wsJsonClient.js",
+    "./wsJsonClientImpl": "./dist/client/wsJsonClientImpl.js",
     "./alertTypes": "./dist/client/types/alertTypes.js",
     "./tdaWsJsonTypes": "./dist/client/tdaWsJsonTypes.js",
     "./wsJsonClientAuth": "./dist/client/wsJsonClientAuth.js",
